Update tests for patientList
