+++ src/containers/PatientList/__tests__/PatientsList.practitioner.test.ts
@@ -1,0 +1,90 @@
+import { renderHook, waitFor } from '@testing-library/react';
+import { SearchParams } from 'fhir-react';
+
+import { formatFHIRDate } from 'aidbox-react/lib/utils/date';
+
+import { isFailure, isSuccess } from 'fhir-react/lib/libs/remoteData';
+
+import { useSearchBar } from 'src/components/SearchBar/hooks';
+import { StringTypeColumnFilterValue } from 'src/components/SearchBar/types';
+
+import { initialSetup } from './utils';
+import { usePatientList } from '../hooks';
+import { getPatientSearchParamsForPractitioner } from '../utils';
+
+async function renderPatientsHooks(searchParams: SearchParams) {
+    const { result } = renderHook(() => {
+        const { columnsFilterValues, onChangeColumnFilter, onResetFilters } = useSearchBar({
+            columns: [
+                {
+                    id: 'patient',
+                    type: 'string',
+                    placeholder: `Search by patient`,
+                },
+            ],
+        });
+
+        const { patientsResponse } = usePatientList(columnsFilterValues as StringTypeColumnFilterValue[], searchParams);
+
+        return {
+            columnsFilterValues,
+            patientsResponse,
+            onChangeColumnFilter,
+            onResetFilters,
+        };
+    });
+
+    return result;
+}
+
+describe('Patient list get by consent', () => {
+    test('Get patients with signed consent', async () => {
+        const data = await initialSetup();
+        const searchParams = getPatientSearchParamsForPractitioner(data.practitioner.id);
+        const result = await renderPatientsHooks(searchParams);
+
+        await waitFor(
+            () => {
+                expect(isSuccess(result.current.patientsResponse)).toBeTruthy();
+            },
+            { timeout: 30000 },
+        );
+        if (isSuccess(result.current.patientsResponse)) {
+            expect(result.current.patientsResponse.data.length).toEqual(1);
+            expect(result.current.patientsResponse.data?.[0]?.id).toEqual(data.patients[2]?.id);
+        }
+    });
+    test('Get patients with incorrect search param status', async () => {
+        const data = await initialSetup();
+        const result = await renderPatientsHooks({ ...data.correctSearchParams, ...{ status: 'draft' } });
+
+        await waitFor(
+            () => {
+                expect(isFailure(result.current.patientsResponse)).toBeTruthy();
+            },
+            { timeout: 30000 },
+        );
+    });
+    test('Get patients with incorrect search param category', async () => {
+        const data = await initialSetup();
+        const result = await renderPatientsHooks({ ...data.correctSearchParams, ...{ category: 'not-data-sharing' } });
+
+        await waitFor(
+            () => {
+                expect(isFailure(result.current.patientsResponse)).toBeTruthy();
+            },
+            { timeout: 30000 },
+        );
+    });
+    test('Get patients with incorrect search param category', async () => {
+        const data = await initialSetup();
+        const result = await renderPatientsHooks({ ...data.correctSearchParams, ...{ actor: 'some-actor-id' } });
+
+        await waitFor(
+            () => {
+                expect(isFailure(result.current.patientsResponse)).toBeTruthy();
+            },
+            { timeout: 30000 },
+        );
+    });
+});
