+++ src/containers/PatientList/__tests__/PatientsList.test.ts
@@ -1,23 +1,12 @@
 import { act, renderHook, waitFor } from '@testing-library/react';
-import { Consent, Period, Reference } from 'fhir/r4b';
 
-import { service } from 'aidbox-react/lib/services/service';
-
-import { useService } from 'fhir-react/lib/hooks/service';
 import { isLoading, isSuccess } from 'fhir-react/lib/libs/remoteData';
-import { createFHIRResource, getReference } from 'fhir-react/lib/services/fhir';
-import { ensure, withRootAccess } from 'fhir-react/lib/utils/tests';
-import { uuid4 } from 'fhir-react/lib/utils/uuid';
-
-import config from 'shared/src/config';
-import { Role, User } from 'shared/src/contrib/aidbox';
 
 import { useSearchBar } from 'src/components/SearchBar/hooks';
 import { StringTypeColumnFilterValue } from 'src/components/SearchBar/types';
-import { createConsent, createPatient, createPractitioner, ensureSave, login, loginAdminUser } from 'src/setupTests';
+import { createPatient, loginAdminUser } from 'src/setupTests';
 
 import { usePatientList } from '../hooks';
-import { getPatientSearchParamsForPractitioner } from '../utils';
 
 const PATIENTS_ADDITION_DATA = [
     {
@@ -144,331 +133,5 @@ describe('Patient list filters testing', () => {
         if (isSuccess(result.current.patientsResponse)) {
             expect(result.current.patientsResponse.data.length).toEqual(2);
         }
-    });
-});
-
-function createConsentData(
-    patientRef: Reference,
-    practitionerRef: Reference,
-    statusCode: 'active' | 'proposed',
-    categoryCode: string,
-    period: Period,
-): Partial<Consent> {
-    return {
-        patient: patientRef,
-        category: [
-            {
-                coding: [
-                    {
-                        code: categoryCode,
-                        system: 'http://terminology.hl7.org/CodeSystem/v3-ActCode',
-                    },
-                ],
-            },
-        ],
-        provision: {
-            type: 'permit',
-            actor: [
-                {
-                    role: {
-                        coding: [
-                            {
-                                code: 'PROV',
-                                system: 'http://terminology.hl7.org/CodeSystem/v3-RoleClass',
-                            },
-                        ],
-                    },
-                    reference: practitionerRef,
-                },
-            ],
-            action: [
-                {
-                    coding: [
-                        {
-                            code: 'access',
-                            system: 'http://terminology.hl7.org/CodeSystem/consentaction',
-                        },
-                    ],
-                },
-            ],
-            period: period,
-            purpose: [
-                {
-                    code: 'CAREMGT',
-                    system: 'http://terminology.hl7.org/CodeSystem/v3-ActReason',
-                    display: 'care management',
-                },
-            ],
-        },
-        resourceType: 'Consent',
-        scope: {
-            coding: [
-                {
-                    code: 'patient-privacy',
-                    system: 'http://terminology.hl7.org/CodeSystem/consentscope',
-                },
-            ],
-        },
-        status: statusCode,
-        performer: [practitionerRef],
-    };
-}
-
-async function patientsListWithConsentSetup(practitionerId: string) {
-    const data = withRootAccess(async () => {
-        const practitioner1 = await createPractitioner({ id: practitionerId });
-        const practitioner2 = await createPractitioner();
-        const patient1 = await createPatient(PATIENTS_ADDITION_DATA[0]);
-        const patient2 = await createPatient(PATIENTS_ADDITION_DATA[1]);
-        const patient3 = await createPatient(PATIENTS_ADDITION_DATA[2]);
-
-        const consentStatus: { incorrect: 'proposed' | 'active'; correct: 'proposed' | 'active' } = {
-            incorrect: 'proposed',
-            correct: 'active',
-        };
-
-        const consentType = {
-            incorrect: 'no-data-sharing',
-            correct: 'data-sharing',
-        };
-
-        const actor = {
-            incorrect: getReference(practitioner2),
-            correct: getReference(practitioner1),
-        };
-
-        const period = {
-            incorrect: {
-                end: '2010-01-01',
-                start: '2020-01-01',
-            },
-            correct: {
-                end: '2030-01-01',
-                start: '2023-01-01',
-            },
-        };
-
-        const consent1 = await createConsent(
-            createConsentData(
-                getReference(patient1),
-                actor['incorrect'],
-                consentStatus['incorrect'],
-                consentType['incorrect'],
-                period['incorrect'],
-            ),
-        );
-        const consent2 = await createConsent(
-            createConsentData(
-                getReference(patient2),
-                actor['incorrect'],
-                consentStatus['incorrect'],
-                consentType['incorrect'],
-                period['correct'],
-            ),
-        );
-        const consent3 = await createConsent(
-            createConsentData(
-                getReference(patient2),
-                actor['incorrect'],
-                consentStatus['incorrect'],
-                consentType['correct'],
-                period['incorrect'],
-            ),
-        );
-        const consent4 = await createConsent(
-            createConsentData(
-                getReference(patient2),
-                actor['incorrect'],
-                consentStatus['correct'],
-                consentType['incorrect'],
-                period['incorrect'],
-            ),
-        );
-        const consent5 = await createConsent(
-            createConsentData(
-                getReference(patient2),
-                actor['correct'],
-                consentStatus['incorrect'],
-                consentType['incorrect'],
-                period['incorrect'],
-            ),
-        );
-        const consent6 = await createConsent(
-            createConsentData(
-                getReference(patient3),
-                actor['correct'],
-                consentStatus['correct'],
-                consentType['correct'],
-                period['correct'],
-            ),
-        );
-
-        return {
-            consents: [consent1, consent2, consent3, consent4, consent5, consent6],
-            patients: [patient1, patient2, patient3],
-        };
-    });
-
-    return data;
-}
-
-describe('Patient list get by consent', () => {
-    const practitionerId = uuid4();
-    beforeAll(async () => {
-        await loginAdminUser();
-    });
-
-    test('Get patients with signed consent', async () => {
-        const data = await patientsListWithConsentSetup(practitionerId);
-
-        const { result } = renderHook(() => {
-            const { columnsFilterValues, onChangeColumnFilter, onResetFilters } = useSearchBar({
-                columns: [
-                    {
-                        id: 'patient',
-                        type: 'string',
-                        placeholder: `Search by patient`,
-                    },
-                ],
-            });
-
-            const searchParams = getPatientSearchParamsForPractitioner(practitionerId);
-
-            const { patientsResponse } = usePatientList(
-                columnsFilterValues as StringTypeColumnFilterValue[],
-                searchParams,
-            );
-
-            return {
-                columnsFilterValues,
-                patientsResponse,
-                onChangeColumnFilter,
-                onResetFilters,
-            };
-        });
-
-        await waitFor(
-            () => {
-                expect(isSuccess(result.current.patientsResponse)).toBeTruthy();
-            },
-            { timeout: 30000 },
-        );
-        if (isSuccess(result.current.patientsResponse)) {
-            expect(result.current.patientsResponse.data.length).toEqual(1);
-            expect(result.current.patientsResponse.data?.[0]?.id).toEqual(data.patients[2]?.id);
-        }
-    });
-});
-
-export async function createUser(userData: Partial<User>) {
-    return ensure(
-        await createFHIRResource<User>({
-            resourceType: 'User',
-            ...userData,
-        }),
-    );
-}
-
-export async function createRole(roleData: Role) {
-    return ensure(
-        await createFHIRResource<Role>({
-            ...roleData,
-        }),
-    );
-}
-
-function practitionerAccessPolicySetup() {
-    const data = withRootAccess(async () => {
-        const practitioner = await createPractitioner();
-        const user = await createUser({
-            fhirUser: {
-                id: practitioner.id!,
-                resourceType: 'Practitioner',
-            },
-            password: 'password',
-            resourceType: 'User',
-            email: 'test@beda.software',
-        });
-        const role = await ensureSave({
-            name: 'practitioner',
-            user: {
-                id: user.id!,
-                resourceType: 'User',
-            },
-            links: {
-                practitioner: {
-                    id: practitioner.id,
-                    resourceType: 'Practitioner',
-                },
-            },
-            resourceType: 'Role',
-        });
-
-        return {
-            practitioner,
-            user,
-            role,
-        };
-    });
-
-    return data;
-}
-
-describe('Access policy', () => {
-    const initialSetup = async () => {
-        const data = await practitionerAccessPolicySetup();
-        await login({ ...data.user, password: 'password' });
-
-        return data;
-    };
-
-    const getResult = async (url: string) => {
-        const { result } = renderHook(() => {
-            const [response] = useService<Consent>(async () =>
-                service({
-                    baseURL: config.baseURL,
-                    method: 'GET',
-                    url: url,
-                }),
-            );
-
-            return response;
-        });
-
-        return result;
-    };
-
-    test.skip('Practitioner can get consent list where practitioner is actor', async () => {
-        const data = await initialSetup();
-        const result = await getResult(`/Consent?actor=${data.practitioner.id}`);
-
-        await waitFor(
-            () => {
-                expect(result.current.status).toEqual('Success');
-            },
-            { timeout: 30000 },
-        );
-    });
-    test.skip('Practitioner cant get consent list with incorrect actor parameter', async () => {
-        await initialSetup();
-        const result = await getResult('/Consent?actor=unknown-practitioner');
-
-        await waitFor(
-            () => {
-                expect(result.current.status).toEqual('Failure');
-            },
-            { timeout: 30000 },
-        );
-    });
-    test.skip('Practitioner cant get consent list without actor parameter', async () => {
-        await initialSetup();
-        const result = await getResult('/Consent');
-
-        await waitFor(
-            () => {
-                expect(result.current.status).toEqual('Failure');
-            },
-            { timeout: 30000 },
-        );
     });
 });
